Added features and bugs
Added some half baked features like editing. this at the moment is simply a button, no functionality.

Started work on a system for the user to make their own query but ran out of time to get it done

some of the code now has been commented, making it a little more readable.